packages/vca_infra/pyproject.toml
[project]
name = "vca-infra"
version = "0.1.0"
requires-python = ">=3.11,<4.0"
dependencies = [
    "sqlmodel",
    "psycopg2-binary",
    "alembic",
    "pydantic-settings",
    "vca-core",
    "google-cloud-storage>=3.7.0",
]

[tool.uv.sources]
vca-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vca_infra"]
